ActiveModel Acceptance Validator is patched after active model is loaded
This version works for rails versions that supports
ActiveSupport.on_load(:active_model) hook and the ones that don't use
it.

The reason to patch this validator is that for some reason it does not
work with reform and to keep it simple I've removed the setup! process
defined in the original code and keep the validator simple.
